Step 125 Project 5 Submission
Minimum requirements include the following:

Utilize the document.write() method and typeof operator to display the data type of a variable of your choosing.

Write an expression combining a string and a number.

Write code utilizing the following operators: ==, ===, >, <, &&, ||, ! and display them in the browser.

Add comments throughout your JavaScript explaining your code.