Trouble with this code.
On line 114 I created an object. When I went to run it nothing happened. I used the console in the google dev tools and it indicated there was a problem with the earlier function on line 104. 
What doesn't make sense is that this function was working before I tried to do the second function 
What is happening here. 
Please also note I have "defer" in my script tag.