jQuery is again not working
I removed the line and the code worked, but then I made some other changes as I was trying to do the other exercises on the same document and it messed up my code again. I took out the additions and it still doesn't work. I am not sure why ??
I don't know why it is not working now.